Enum rusttype::CodepointOrGlyphId

pub enum CodepointOrGlyphId {
    Codepoint(Codepoint),
    GlyphId(GlyphId),
}

Represents either a Unicode code point, or a glyph identifier for a font.

This is used as input for functions that can accept code points or glyph identifiers.

You typically won't construct this type directly, instead relying on From and Into.

Variants
Codepoint(Codepoint)GlyphId(GlyphId)