Struct rusttype::Glyph

pub struct Glyph<'a> { /* fields omitted */ }

A single glyph of a font. this may either be a thin wrapper referring to the font and the glyph id, or it may be a standalone glyph that owns the data needed by it.

A Glyph does not have an inherent scale or position associated with it. To augment a glyph with a size, give it a scale using scaled. You can then position it using positioned.

Methods

impl<'a> Glyph<'a>

fn font(&self) -> Option<&Font<'a>>

The font to which this glyph belongs. If the glyph is a standalone glyph that owns its resources, it no longer has a reference to the font which it was created from (using standalone()). In which case, None is returned.

fn id(&self) -> GlyphId

The glyph identifier for this glyph.

fn scaled(self, scale: Scale) -> ScaledGlyph<'a>

Augments this glyph with scaling information, making methods that depend on the scale of the glyph available.

fn standalone(&self) -> Glyph<'static>

Turns a Glyph<'a> into a Glyph<'static>. This produces a glyph that owns its resources, extracted from the font. This glyph can outlive the font that it comes from.

Calling standalone() on a standalone glyph shares the resources, and is equivalent to clone().