Struct rusttype::Scale

pub struct Scale(_, _);

An opaque struct representing a common format for font scaling. You typically won't use this struct directly, instead using Pixels or PixelsXY and the Into trait to pass them to functions.

You can however implement your own scales for use with functions that accept S: Into<Scale>. As a simplified example, if you want to write scales in a different unit, like inches, and you know that there are 96 pixels in an inch in your use case, you can create a struct like the following:

use rusttype::{Pixels, Scale};

struct Inches(f32);
impl From<Inches> for Scale {
    fn from(i: Inches) -> Scale {
        Pixels(i.0 * 96.0).into()
    }
}

You can then use Inches wherever you could use Pixels or PixelsXY before.