Struct rusttype::ScaledGlyph

pub struct ScaledGlyph<'a> { /* fields omitted */ }

A glyph augmented with scaling information. You can query such a glyph for information that depends on the scale of the glyph.

Implementations

impl<'a> ScaledGlyph<'a>

pub fn id(&self) -> GlyphId

The glyph identifier for this glyph.

pub fn font(&self) -> Option<&Font<'a>>

The font to which this glyph belongs. If the glyph is a standalone glyph that owns its resources, it no longer has a reference to the font which it was created from (using standalone()). In which case, None is returned.

pub fn into_unscaled(self) -> Glyph<'a>

A reference to this glyph without the scaling

pub fn unscaled(&self) -> &Glyph<'a>

Removes the scaling from this glyph

pub fn positioned(self, p: Point<f32>) -> PositionedGlyph<'a>

Augments this glyph with positioning information, making methods that depend on the position of the glyph available.

pub fn scale(&self) -> Scale

pub fn h_metrics(&self) -> HMetrics

Retrieves the "horizontal metrics" of this glyph. See HMetrics for more detail.

pub fn shape(&self) -> Option<Vec<Contour>>

Produces a list of the contours that make up the shape of this glyph. Each contour consists of a sequence of segments. Each segment is either a straight Line or a Curve.

The winding of the produced contours is clockwise for closed shapes, anticlockwise for holes.

pub fn exact_bounding_box(&self) -> Option<Rect<f32>>

The bounding box of the shape of this glyph, not to be confused with pixel_bounding_box, the conservative pixel-boundary bounding box. The coordinates are relative to the glyph's origin.

pub fn standalone(&self) -> ScaledGlyph<'static>

Constructs a glyph that owns its data from this glyph. This is similar to Glyph::standalone. See that function for more details.