Struct rusttype::SizedGlyph

pub struct SizedGlyph<'a> {
    // some fields omitted
}

A glyph augmented with scaling information. You can query such a glyph for information that depends on the scale of the glyph.

Methods

impl<'a> SizedGlyph<'a>

fn positioned(self, p: Point<f32>) -> PositionedGlyph<'a>

Augments this glyph with positioning information, making methods that depend on the position of the glyph available.

fn h_metrics(&self) -> HMetrics

Retrieves the "horizontal metrics" of this glyph. See HMetrics for more detail.

fn shape(&self) -> Option<Vec<Contour>>

Produces a list of the contours that make up the shape of this glyph. Each contour consists of a sequence of segments. Each segment is either a straight Line or a Curve.

The winding of the produced contours is clockwise for closed shapes, anticlockwise for holes.

fn exact_bounding_box(&self) -> Option<Rect<f32>>

The bounding box of the shape of this glyph, not to be confused with pixel_bounding_box, the conservative pixel-boundary bounding box. The coordinates are relative to the glyph's origin.

fn standalone(&self) -> SizedGlyph<'static>

Constructs a glyph that owns its data from this glyph. This is similar to Glyph::standalone. See that function for more details.