rusttype
Trait IntoGlyphId

pub trait IntoGlyphId {
    // Required method
    fn into_glyph_id(self, _: &Font<'_>) -> GlyphId;
}

A trait for types that can be converted into a GlyphId, in the context of a specific font.

Many rusttype functions that operate on characters accept values of any type that implements IntoGlyphId. Such types include char, Codepoint, and obviously GlyphId itself.

Required Methods

fn into_glyph_id(self, _: &Font<'_>) -> GlyphId

Convert self into a GlyphId, consulting the index map of font if necessary.