Struct rusttype::Font

pub struct Font<'a> { /* fields omitted */ }

A single font. This may or may not own the font data.

Methods

impl<'a> Font<'a>

fn v_metrics(&self, scale: Scale) -> VMetrics

The "vertical metrics" for this font at a given scale. These metrics are shared by all of the glyphs in the font. See VMetrics for more detail.

fn glyph_count(&self) -> usize

The number of glyphs present in this font. Glyph identifiers for this font will always be in the range 0..self.glyph_count()

fn glyph<C: Into<CodepointOrGlyphId>>(&self, id: C) -> Option<Glyph>

Returns the corresponding glyph for a Unicode code point or a glyph id for this font. If id corresponds to a glyph identifier, the identifier must be valid (smaller than self.glyph_count()), otherwise None is returned.

Note that code points without corresponding glyphs in this font map to the "undef" glyph, glyph 0.

fn glyphs_for<I: Iterator>(&self, itr: I) -> GlyphIter<I> where I::Item: Into<CodepointOrGlyphId>

A convenience function.

Returns an iterator that produces the glyphs corresponding to the code points or glyph ids produced by the given iterator itr.

This is equivalent in behaviour to itr.map(|c| font.glyph(c).unwrap()).

fn layout<'b, 'c>(&'b self,
s: &'c str,
scale: Scale,
start: Point<f32>)
-> LayoutIter<'b, 'c>

A convenience function for laying out glyphs for a string horizontally. It does not take control characters like line breaks into account, as treatment of these is likely to depend on the application.

Note that this function does not perform Unicode normalisation. Composite characters (such as ö constructed from two code points, ¨ and o), will not be normalised to single code points. So if a font does not contain a glyph for each separate code point, but does contain one for the normalised single code point (which is common), the desired glyph will not be produced, despite being present in the font. Deal with this by performing Unicode normalisation on the input string before passing it to layout. The crate unicode-normalization is perfect for this purpose.

Calling this function is equivalent to a longer sequence of operations involving glyphs_for, e.g.

font.layout("Hello World!", scale, start)

produces an iterator with behaviour equivalent to the following:

font.glyphs_for("Hello World!".chars())
    .scan((None, 0.0), |&mut (mut last, mut x), g| {
        let g = g.scaled(scale);
        let w = g.h_metrics().advance_width
            + last.map(|last| font.pair_kerning(scale, last, g.id())).unwrap_or(0.0);
        let next = g.positioned(start + vector(x, 0.0));
        last = Some(next.id());
        x += w;
        Some(next)
    })

fn pair_kerning<A, B>(&self, scale: Scale, first: A, second: B) -> f32 where A: Into<CodepointOrGlyphId>, B: Into<CodepointOrGlyphId>

Returns additional kerning to apply as well as that given by HMetrics for a particular pair of glyphs.