Struct rusttype::PositionedGlyph

pub struct PositionedGlyph<'font> {
    pub(crate) sg: ScaledGlyph<'font>,
    pub(crate) position: Point<f32>,
    pub(crate) bb: Option<Rect<i32>>,
}

A glyph augmented with positioning and scaling information. You can query such a glyph for information that depends on the scale and position of the glyph.

Fields
sg: ScaledGlyph<'font>position: Point<f32>bb: Option<Rect<i32>>
Implementations

impl<'font> PositionedGlyph<'font>

pub fn id(&self) -> GlyphId

The glyph identifier for this glyph.

pub fn font(&self) -> &Font<'font>

The font to which this glyph belongs.

pub fn unpositioned(&self) -> &ScaledGlyph<'font>

A reference to this glyph without positioning

pub fn into_unpositioned(self) -> ScaledGlyph<'font>

Removes the positioning from this glyph

pub fn pixel_bounding_box(&self) -> Option<Rect<i32>>

The conservative pixel-boundary bounding box for this glyph. This is the smallest rectangle aligned to pixel boundaries that encloses the shape of this glyph at this position. Note that the origin of the glyph, at pixel-space coordinates (0, 0), is at the top left of the bounding box.

pub fn scale(&self) -> Scale

pub fn position(&self) -> Point<f32>

pub fn build_outline(&self, builder: &mut impl OutlineBuilder) -> bool

Builds the outline of the glyph with the builder specified. Returns false when the outline is either malformed or empty.

pub fn draw<O: FnMut(u32, u32, f32)>(&self, o: O)

Rasterises this glyph. For each pixel in the rect given by pixel_bounding_box(), o is called:

o(x, y, v)

where x and y are the coordinates of the pixel relative to the min coordinates of the bounding box, and v is the analytically calculated coverage of the pixel by the shape of the glyph. Calls to o proceed in horizontal scanline order, similar to this pseudo-code:

let bb = glyph.pixel_bounding_box();
for y in 0..bb.height() {
    for x in 0..bb.width() {
        o(x, y, calc_coverage(&glyph, x, y));
    }
}

pub fn set_position(&mut self, p: Point<f32>)

Resets positioning information and recalculates the pixel bounding box