Struct rusttype::PositionedGlyph

pub struct PositionedGlyph<'a> {
    // some fields omitted
}

A glyph augmented with positioning and scaling information. You can query such a glyph for information that depends on the scale and position of the glyph.

Methods

impl<'a> PositionedGlyph<'a>

fn pixel_bounding_box(&self) -> Option<Rect<i32>>

The conservative pixel-boundary bounding box for this glyph. This is the smallest rectangle aligned to pixel boundaries that encloses the shape of this glyph at this position.

fn positioned_shape(&self) -> Option<Vec<Contour>>

Similar to shape(), but with the position of the glyph taken into account.

fn draw<O: FnMut(u32, u32, f32)>(&self, o: O)

Rasterises this glyph. For each pixel in the rect given by pixel_bounding_box(), o is called:

o(x, y, v)

where x and y are the coordinates of the pixel relative to the min coordinates of the bounding box, and v is the analytically calculated coverage of the pixel by the shape of the glyph. Calls to o proceed in horizontal scanline order, similar to this pseudo-code:

let bb = glyph.pixel_bounding_box();
for y in 0..bb.height() {
    for x in 0..bb.width() {
        o(x, y, calc_coverage(&glyph, x, y));
    }
}

fn standalone(&self) -> PositionedGlyph<'static>

Constructs a glyph that owns its data from this glyph. This is similar to Glyph::standalone. See that function for more details.